+++ genesyscloud/resource_genesyscloud_init.go
@@ -15,7 +15,6 @@ func registerDataSources(l registrar.Registrar) {
 
 	l.RegisterDataSource("genesyscloud_auth_division_home", DataSourceAuthDivisionHome())
 	l.RegisterDataSource("genesyscloud_journey_outcome", dataSourceJourneyOutcome())
-	l.RegisterDataSource("genesyscloud_journey_segment", dataSourceJourneySegment())
 	l.RegisterDataSource("genesyscloud_knowledge_knowledgebase", dataSourceKnowledgeKnowledgebase())
 	l.RegisterDataSource("genesyscloud_organizations_me", DataSourceOrganizationsMe())
 	l.RegisterDataSource("genesyscloud_quality_forms_evaluation", DataSourceQualityFormsEvaluations())
@@ -26,7 +25,6 @@ func registerDataSources(l registrar.Registrar) {
 func registerResources(l registrar.Registrar) {
 
 	l.RegisterResource("genesyscloud_journey_outcome", ResourceJourneyOutcome())
-	l.RegisterResource("genesyscloud_journey_segment", ResourceJourneySegment())
 	l.RegisterResource("genesyscloud_knowledge_knowledgebase", ResourceKnowledgeKnowledgebase())
 	l.RegisterResource("genesyscloud_quality_forms_evaluation", ResourceEvaluationForm())
 	l.RegisterResource("genesyscloud_quality_forms_survey", ResourceSurveyForm())
@@ -36,7 +34,6 @@ func registerResources(l registrar.Registrar) {
 
 func registerExporters(l registrar.Registrar) {
 	l.RegisterExporter("genesyscloud_journey_outcome", JourneyOutcomeExporter())
-	l.RegisterExporter("genesyscloud_journey_segment", JourneySegmentExporter())
 	l.RegisterExporter("genesyscloud_knowledge_knowledgebase", KnowledgeKnowledgebaseExporter())
 	l.RegisterExporter("genesyscloud_quality_forms_evaluation", EvaluationFormExporter())
 	l.RegisterExporter("genesyscloud_quality_forms_survey", SurveyFormExporter())
